Reorganize the src folder
