+++ src/pages/home.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import OverlayMenu from '../components/home/OverlayMenu';
 import Panel from '../components/home/Panel';
-import useSideMenu from '../hooks/sideMenu';
+import useSideMenu from '../states/sideMenu';
 import UserDisplay from '../components/home/UserDisplay';
 
 const Home = () => {
