+++ src/pages/register.tsx
@@ -3,7 +3,7 @@ import { PropagateLoader } from 'react-spinners';
 import CenterCard from '../components/register/CenterCard';
 import RegisterForm from '../components/register/RegisterForm';
 import ServerModal from '../components/register/ServerModal';
-import useVersion from '../hooks/version';
+import useVersion from '../states/getVersion';
 
 const Register = () => {
   const { isLoading, error, data } = useVersion();
