+++ src/states/user/renameUser.ts
@@ -1,7 +1,7 @@
 import { useContext } from 'react';
 import { useMutation } from 'react-query';
-import SettingsContext from '../contexts/SettingContext';
-import UserContext from '../contexts/UserContext';
+import { SettingsContext } from '../settings';
+import { UserContext } from './user';
 
 const useRenameUser = (name: string) => {
   const settings = useContext(SettingsContext);
